Se agregan correciones

+++ Intelificio-Back/Features/Community/Commands/Delete/DeleteCommunityCommandHandler.cs
@@ -19,8 +19,10 @@ namespace Backend.Features.Community.Commands.Delete
 
         public async Task<Result> Handle(DeleteCommunityCommand request, CancellationToken cancellationToken)
         {
-            var community = await _context.Community.FirstOrDefaultAsync(x => x.ID == request.Id);
+            var community = await _context.Community.Include(x => x.Buildings).FirstOrDefaultAsync(x => x.ID == request.Id);
             if (community == null) return Result.Failure(CommunityErrors.CommunityNotFoundDelete);
+
+            if (community.Buildings.Count >= 1) return Result.Failure(CommunityErrors.HasAssignedBuildingsOnDelete);
             _ = _context.Community.Remove(community);
             _ = await _context.SaveChangesAsync();
             return Result.Success();
